extract `EditCommentTextState` and `EditCommentPreviewerState`
